Fixed Color data types

+++ Sources/LedStrip/Color.swift
@@ -1,8 +1,8 @@
 
 public struct Color {
-	var red: UInt8
-	var green: UInt8
-	var blue: UInt8
+	var red: UInt32
+	var green: UInt32
+	var blue: UInt32
 
 	public static let white = Self(red: 255, green: 255, blue: 255)
 	public static let gray = Self(red: 128, green: 128, blue: 128)
